Convert text elements to svg
This gives us the ability to add a stroke/outline outside of the text (the text-stroke css on regular text is only supported as a -webkit- extension, and even then it strokes inside the text, not outside).

The original goal here was to allow chroma-keying the background on OBS but still having the text readable, but after I finished this I realized that the background changing color when we go over time kinda blows that up. So this can be reverted if needed, but I want the commit in history so I know how to do it.